[Backend] rename mvc arquitecture cart v1 - from domain to model, create new agregate root, entities, vo, and exceptions, handler business rule quantity
