+++ services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/cart/domain/model/exception/CartItemNotFoundDomainException.java
@@ -1,0 +1,9 @@
+package org.ecommerce.blackfriday.cart.domain.model.exception;
+
+import org.ecommerce.blackfriday.common.exception.DomainException;
+
+public class CartItemNotFoundDomainException extends DomainException {
+    public CartItemNotFoundDomainException(String message) {
+        super(message);
+    }
+}
