+++ services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/cart/domain/model/exception/InvalidQuantityDomainException.java
@@ -1,0 +1,9 @@
+package org.ecommerce.blackfriday.cart.domain.model.exception;
+
+import org.ecommerce.blackfriday.common.exception.DomainException;
+
+public class InvalidQuantityDomainException extends DomainException {
+    public InvalidQuantityDomainException(String message) {
+        super(message);
+    }
+}
