+++ services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/cart/domain/model/factory/CartFactory.java
@@ -1,0 +1,4 @@
+package org.ecommerce.blackfriday.cart.domain.model.factory;
+
+public class CartFactory {
+}
