+++ services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/cart/domain/model/valueobject/CartId.java
@@ -1,0 +1,11 @@
+package org.ecommerce.blackfriday.cart.domain.model.valueobject;
+
+import org.ecommerce.blackfriday.common.domain.model.valueobject.BaseId;
+
+import java.util.UUID;
+
+public class CartId extends BaseId<UUID> {
+    protected CartId(UUID value) {
+        super(value);
+    }
+}
