+++ services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/cart/domain/model/valueobject/CustomerId.java
@@ -1,0 +1,12 @@
+package org.ecommerce.blackfriday.cart.domain.model.valueobject;
+
+import org.ecommerce.blackfriday.common.domain.model.valueobject.BaseId;
+
+import java.util.UUID;
+
+public class CustomerId extends BaseId<UUID> {
+
+    protected CustomerId(UUID value) {
+        super(value);
+    }
+}
