+++ services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/cart/domain/model/valueobject/Quantity.java
@@ -1,0 +1,49 @@
+package org.ecommerce.blackfriday.cart.domain.model.valueobject;
+
+import org.ecommerce.blackfriday.cart.domain.model.exception.InvalidQuantityDomainException;
+
+import java.util.Objects;
+
+public class Quantity {
+
+    private final int value;
+
+    public Quantity(int value) {
+        if (value < 1) {
+            throw new InvalidQuantityDomainException("Invalid Quantity, must be greater to ZERO");
+        }
+
+        this.value = value;
+    }
+
+    public int getValue() {
+        return value;
+    }
+
+    public Quantity decrement () {
+        if (this.value == 1)
+            throw new InvalidQuantityDomainException("Invalid operation, It cannot decrement below ONE (1)");
+
+        return new Quantity(this.value - 1);
+    }
+
+    public Quantity increment () {
+        return new Quantity(this.value + 1);
+    }
+
+    public Quantity set(int newValue) {
+        return new Quantity(newValue);
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (o == null || getClass() != o.getClass()) return false;
+        Quantity quantity = (Quantity) o;
+        return value == quantity.value;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hashCode(value);
+    }
+}
