+++ services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/cart/exceptions/ApiError.java
@@ -1,0 +1,4 @@
+package org.ecommerce.blackfriday.cart.exceptions;
+
+public record ApiError(int status, String message) {
+}
