+++ services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/cart/exceptions/CartItemNotFoundException.java
@@ -1,0 +1,8 @@
+package org.ecommerce.blackfriday.cart.exceptions;
+
+public class CartItemNotFoundException extends RuntimeException {
+
+    public CartItemNotFoundException(String cartId, String itemId) {
+        super("Cart with the [ID] " + cartId + " does not have this item added with the [ID] " + itemId);
+    }
+}
