+++ services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/product/managment/domain/exception/InvalidProductNameDomainException.java
@@ -1,0 +1,10 @@
+package org.ecommerce.blackfriday.product.managment.domain.exception;
+
+import org.ecommerce.blackfriday.common.exception.DomainException;
+
+public class InvalidProductNameDomainException extends DomainException {
+
+    public InvalidProductNameDomainException(String message) {
+        super(message);
+    }
+}
