+++ services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/product/managment/domain/valueobject/ProductName.java
@@ -1,17 +1,19 @@
 package org.ecommerce.blackfriday.product.managment.domain.valueobject;
 
+import org.ecommerce.blackfriday.product.managment.domain.exception.InvalidProductNameDomainException;
+
 import java.util.Objects;
 
 public record ProductName (String name) {
 
     public void validateProductName () {
         if (Objects.isNull(this.name) || this.name.isBlank()) {
-            throw new IllegalArgumentException("Product Name is required");
+            throw new InvalidProductNameDomainException("Product Name is required");
         }
 
         final int CANT_MIN_CHARACTERS = 3;
         if (this.name.length() < CANT_MIN_CHARACTERS) {
-            throw new IllegalArgumentException("Product Name is invalid, must have min 5 characters");
+            throw new InvalidProductNameDomainException("Product Name is invalid, must have min 5 characters");
         }
     }
