--- services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/shopping/cart/exceptions/ApiError.java
@@ -1,4 +1,0 @@
-package org.ecommerce.blackfriday.shopping.cart.exceptions;
-
-public record ApiError(int status, String message) {
-}
