--- services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/shopping/cart/exceptions/CartNotFoundException.java
@@ -1,8 +1,0 @@
-package org.ecommerce.blackfriday.shopping.cart.exceptions;
-
-public class CartNotFoundException extends RuntimeException {
-
-    public CartNotFoundException(String cartId) {
-        super("Cart with the [ID] " + cartId + " not found.");
-    }
-}
