--- services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/shopping/cart/exceptions/ShoppingCartGlobalExceptionHandler.java
@@ -1,23 +1,0 @@
-package org.ecommerce.blackfriday.shopping.cart.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class ShoppingCartGlobalExceptionHandler {
-
-    @ExceptionHandler(CartNotFoundException.class)
-    public ResponseEntity<ApiError> handlerCartNotFound (CartNotFoundException ex) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                .body(new ApiError(HttpStatus.NOT_FOUND.value(), ex.getMessage()));
-    }
-
-    @ExceptionHandler(CartItemNotFoundException.class)
-    public ResponseEntity<ApiError> handlerCartItemNotFound (CartItemNotFoundException ex) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                .body(new ApiError(HttpStatus.NOT_FOUND.value(), ex.getMessage()));
-    }
-
-}
